refactor: rename to workflows dir
