+++ packages/cli/src/workflows/setupProject.ts
